Rename client -> client_core; copter_client -> client
